requirements.txt
APScheduler==3.9.1
attrs==21.4.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
blinker==1.4
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.2
Deprecated==1.2.13
dirhash==0.2.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Limiter==2.4.6
Flask-Mail==0.9.0
flask-mailman==0.3.0
Flask-MySQLdb==1.0.1
flask-talisman==1.0.0
Flask-WTF==1.0.1
google-api-core==2.8.1
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
limits==2.6.3
MarkupSafe==2.1.1
mkdocs-material-extensions==1.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
protobuf==3.20.1
py-healthcheck==1.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyotp==2.6.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.8
scandir==1.10.0
scantree==0.0.1
six==1.16.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
